Add live status polling to frontend. And a few others:
Migrate from codesearch_ui to codesearch.js
Some comment cleanup in a few places
Remove some unused imports
