+++ web/cmd/build-web/main.go
@@ -6,16 +6,8 @@ import (
 	"flag"
 	"os"
 	"path"
-	"regexp"
 
 	"github.com/evanw/esbuild/pkg/api"
-	"github.com/tdewolff/minify/v2"
-	"github.com/tdewolff/minify/v2/css"
-	"github.com/tdewolff/minify/v2/html"
-	"github.com/tdewolff/minify/v2/js"
-	// "github.com/tdewolff/minify/v2/json"
-	// "github.com/tdewolff/minify/v2/svg"
-	// "github.com/tdewolff/minify/v2/xml"
 )
 
 func main() {
@@ -25,11 +17,6 @@ func main() {
 	// bazel-out/darwin_arm64-fastbuild/bin/web
 	outDir := flag.String("test", "", "directory to emit files to")
 	flag.Parse()
-
-	m := minify.New()
-	m.AddFunc("text/css", css.Minify)
-	m.AddFunc("text/html", html.Minify)
-	m.AddFuncRegexp(regexp.MustCompile("^(application|text)/(x-)?(java|ecma)script$"), js.Minify)
 
 	// build JS
 	result := api.Build(api.BuildOptions{
