+++ web/src/codesearch/codesearch.js
@@ -1,33 +1,27 @@
-$ = require('jquery');
-_ = require('underscore');
+function getSelectedText() {
+  return window.getSelection ? window.getSelection().toString() : null;
+}
 
-"use strict";
+var searchBox;
+var errorsBox;
+var resultsContainer;
+var helpArea;
+var caseSelect;
+var regexToggle;
 
+var searchResults; // giant HTML string
 var liveDot;
 var liveText;
-var liveSubInfo; 
 
-var Codesearch = function() {
-  return {
-    delegate: null,
-    retry_time: 50,
-    next_search: null,
-    in_flight: null,
-    connect: function(delegate) {
-      if (delegate !== undefined)
-        Codesearch.delegate = delegate;
-      if (Codesearch.delegate.on_connect)
-        setTimeout(Codesearch.delegate.on_connect, 0)
-      liveDot = document.getElementById("live-status-dot");
-      liveText = document.getElementById("live-status-text");
-      liveSubInfo = document.getElementById("live-status-sub-info");
-    },
-    live_poll: function() {
+var searchOptions = {
+  q: '',
+  regex: false,
+  context: true, // we don't have an option for disabling context. No one uses it
+  case: 'auto',
+}
+
+function livePoll() {
       var url = "/api/v1/bkstatus/";
-      var lastText = "";
-      var lastDate;
-      var lastDateNum;
-      // TODO: add the selected bk from the select
       setInterval(function () {
         // Don't make network requests while tab is in background
         if (document.hidden) return;
@@ -46,26 +40,19 @@ var Codesearch = function() {
               // state, then add a "reload" button
 
               liveDot.dataset.status = 'up';
-              liveText.innerText = 'Connected';
-
-              var indexText = 'Querying index created ' + split[1] + ' ago';
-              if (indexText !== liveSubInfo.innerText) {
-                liveSubInfo.innerText = indexText;
-              }
+              liveText.innerText = 'Connected. Index age: ' + split[1];
             } else {
               var timeDown = split[1];
 
               // https://pkg.go.dev/google.golang.org/grpc@v1.45.0/codes#Code
               if (status === '14') {
                 liveDot.dataset.status = 'reloading';
-                liveText.innerText = 'Search Index Reloading (' + timeDown + ')';
-                liveSubInfo.innerText = 'Queries will resolve once the index is reloaded';
+                liveText.innerText = 'Index reloading.. (' + timeDown + ')';
                 return;
               }
               // Some other unhandled grpc issue
               liveDot.dataset.status = 'down';
-              liveText.innerText = 'Disconnected (' + timeDown + ')';
-              liveSubInfo.innerText = 'Search backend down'
+              liveText.innerText = 'Disconnected. (' + timeDown + ')';
             }
          })
         .catch(function (error) {
@@ -75,69 +62,249 @@ var Codesearch = function() {
           liveSubInfo.innerText = 'No connection to webserver';
         });
       }, 1000);
-    },
-    new_search: function(opts) {
-      Codesearch.next_search = opts;
-      if (Codesearch.in_flight == null)
-        Codesearch.dispatch()
-    },
-    dispatch: function() {
-      if (!Codesearch.next_search)
-        return;
-      Codesearch.in_flight = Codesearch.next_search;
-      Codesearch.next_search = null;
+}
 
-      var opts = Codesearch.in_flight;
+var currUrl;
+// We could maybe rethink this to be a function that only updates
+// the searchapram for the changed option. Ok for now tho
+function updateSearchParamState() {
+  if (!currUrl) {
+    url = new URL(window.location);
+  }
 
-      var url = "/api/v1/search/";
-      if ('backend' in opts) {
-        url = url + opts.backend;
-      }
-      var q = {
-        q: opts.q,
-        fold_case: opts.fold_case,
-        regex: opts.regex
-      };
+  var sp = url.searchParams;
 
-      url = url + "?" + $.param(q);
+  sp.set('q', encodeURIComponent(searchOptions.q));
+  sp.set('regex', searchOptions.regex);
+  sp.set('fold_case', searchOptions.case);
+  window.history.pushState({}, '', url);
 
-      var xhr = $.getJSON(url);
-      var start = new Date();
-      xhr.done(function (data) {
-        var elapsed = new Date() - start;
-        data.results.forEach(function (r) {
-          Codesearch.delegate.match(opts.id, r);
-        });
-        data.file_results.forEach(function (r) {
-          Codesearch.delegate.file_match(opts.id, r);
-        });
-        data.tree_results.forEach(function (t) {
-          Codesearch.delegate.tree_match(opts.id, t);
-        });
-        Codesearch.delegate.search_done(opts.id, elapsed, data.search_type, data.info.why);
-      });
-      xhr.fail(function(data) {
-        window._err = data;
-        if (data.status >= 400 && data.status < 500) {
-          var err = JSON.parse(data.responseText);
-          Codesearch.delegate.error(opts.id, err.error.message);
-        } else {
-          var message = "Cannot connect to server";
-          if (data.status) {
-            message = "Bad response " + data.status + " from server";
-          }
-          Codesearch.delegate.error(opts.id, message);
-          console.log("server error", data.status, data.responseText);
-        }
-      });
-      xhr.always(function() {
-        Codesearch.in_flight = null;
-        setTimeout(Codesearch.dispatch, 0);
-      });
-    }
+  doSearch();
+}
+
+// Take the present search options, perform a search
+// then update the 
+function doSearch() {
+  if (searchOptions.q === '') {
+    helpArea.style.display = "initial";
+    resultsContainer.innerHTML = "";
+    errorsBox.style.display = "none";
+    return;
   };
-}();
+  var time1 = performance.now();
+  var time2;
+  fetch("/api/v2/getRenderedSearchResults/?q=" + 
+    searchOptions.q + "&fold_case=" + searchOptions.case + "&regex=" + searchOptions.regex + "&context=" + 
+    searchOptions.context)
+  .then(function(r) {
+    time2 = performance.now();
+    if (!r.ok) {
+      return Promise.reject(r.text());
+    } else {
+      return r.text();
+    }
+  })
+  .then(function (text) {
+    helpArea.style.display = "none";
+    resultsContainer.innerHTML = text;
+    errorsBox.style.display = "none";
+
+    var timeTaken = ((time2 - time1) / 1000).toFixed(4);
+    document.getElementById('searchtime').innerText = timeTaken + "s";
+    document.getElementById('searchtimebox').style.display = "initial";
+  })
+  .catch(function (err) {
+    err.then(function (errText) {
+      // display the help area, clear previous results, and show the new error
+      helpArea.style.display="initial";
+      resultsContainer.innerHTML = "";
+      errorsBox.querySelector('#errortext').innerText = errText;
+      errorsBox.style.display = "initial";
+    });
+  });
+}
+
+function updateQuery(inputEvnt) {
+  searchOptions.q = inputEvnt.target.value;
+  updateSearchParamState();
+}
+
+function toggleControlButton() {
+  var currValue = this.getAttribute('data-selected') === 'true';
+  this.setAttribute('data-selected', !currValue);
+  searchOptions[this.getAttribute('name')] = !currValue;
+  updateSearchParamState();
+}
+
+// Set the textInput value and all selection controls
+// TODO: validate the given options
+var validControlOptions = {
+  "regex": [true, false],
+  "context": [true, false],
+  "case": ["auto", false, true]
+};
+function initStateFromQueryParams() {
+  var currURL = new URL(document.location);
+  var sp = currURL.searchParams;
+
+  var currentQ = decodeURIComponent(sp.get('q') || '');
+  var caseVal = sp.get('fold_case') || 'auto'; 
+  var regexVal = sp.get('regex') || false;
+
+  searchBox.value = currentQ;
+  caseSelect.value = caseVal
+  regexToggle.dataset.selected = regexVal; 
+  
+
+  searchOptions = {
+    q: currentQ,
+    regex: regexVal,
+    context: sp.get('context') || true,
+    case: caseVal,
+  };
+
+
+  doSearch();
+}
+
+function renderSearchHistory() {
+  var currHistory = localStorage.getItem('search-history') || '[]';
+      try {
+        currHistory = JSON.parse(currHistory);
+      } catch (err) {
+        console.error('error parsing localStorage search history. Resetting it.');
+        currHistory = [];
+      }
+
+      var c = document.querySelector('#helparea #recent-searches .searches-container');
+      // empty the container. Note we can't replaceChildren(historyElems)
+      // because the current UglifyJs plugin used by webpack doesn't support the
+      // .../spread operator and replaceChildren expects a comma delimited list
+      // of nodes
+      c.replaceChildren();
+
+      for (var i = 0; i < currHistory.length; i++) {
+        var searchText = currHistory[i];
+        var elem = document.createElement('button');
+        elem.innerText = searchText;
+        elem.title = 'Do search for: ' + searchText;
+        elem.classList.add('search-item');
+        elem.addEventListener('click', function(e) {
+          searchBox.value = e.target.innerText;
+          searchBox.dispatchEvent(new Event('input'))
+        });
+        c.appendChild(elem);
+      };
+}
+
+function addSearchQueryToHistory(e) {
+    if (e.target.value.trim() == '') {
+      return
+    };
+
+    var currHistory = localStorage.getItem('search-history') || '[]';
+      try {
+        currHistory = JSON.parse(currHistory);
+      } catch (err) {
+        console.error('error parsing localStorage search history. Resetting it.');
+        currHistory = [];
+      }
+
+    var dedupedHistory = currHistory.filter(function (hElem) {
+        return hElem !== e.target.value;
+      });
+
+    dedupedHistory.unshift(e.target.value); // Add the new item to the front
+    dedupedHistory = dedupedHistory.slice(0, 5); // Only keep the last 5 entries
+
+    localStorage.setItem('search-history', JSON.stringify(dedupedHistory));
+    renderSearchHistory();
+}
+
+function toggleMoreFileMatches(e) {
+  document.querySelector('.path-results .extra-results').classList.toggle('hidden');
+  var textContainer = e.currentTarget.querySelector('#toggle-btn-text');
+  textContainer.innerText = textContainer.innerText === 'Show all' ? 'Show less' : 'Show all';
+  e.currentTarget.querySelector('img').classList.toggle('open');
+}
+
+function handleFileExtBtnClick(e) {
+  var q = searchBox.value;
+  var ext = e.target.innerText;
+  if (regexToggle.dataset.selected == "true") {
+    q = 'path:\\' + ext + '$ ' + q;
+  }
+  else {
+    q = 'path:' + ext + ' ' + q;
+  }
+  searchBox.value = q;
+  searchBox.dispatchEvent(new Event('input'))
+}
+
+function init() {
+  "use strict"
+
+  searchBox = document.getElementById('searchbox')
+  resultsContainer = document.querySelector('#resultarea > #results');
+  helpArea = document.getElementById('helparea');
+  caseSelect = document.getElementById('case-sensitivity-toggle');
+  regexToggle = document.querySelector('button[id=toggle-regex]');
+  errorsBox = document.getElementById('regex-error');
+  liveDot = document.getElementById("live-status-dot");
+  liveText = document.getElementById("live-status-text");
+
+  caseSelect.addEventListener('change', function (e) {
+    var newVal = e.target.value;
+    searchOptions['case'] = newVal;
+    updateSearchParamState();
+  });
+  regexToggle.addEventListener('click', toggleControlButton);
+  searchBox.addEventListener('input', updateQuery);
+
+  // add search events to recent searches
+  searchBox.addEventListener('blur', addSearchQueryToHistory);
+
+  document.addEventListener('click', function(e) {
+    var clickedElem = event.target;
+
+    var btn = e.target.closest('button');
+    if (btn && btn.id == "showMoreFilematchesBtn") {
+      toggleMoreFileMatches(e);
+    } else if (btn && btn.classList.contains("file-extension")) {
+      handleFileExtBtnClick(e);
+    }
+  });
+
+  // listen for the '/' key to trigger search input focus
+  // or, if text is selected, trigger a search for it
+  document.addEventListener('keyup', function (e) {
+     if (event.altKey || event.ctrlKey || event.metaKey || event.shiftKey)
+      return;
+     if (event.key !== "/" || (searchBox === document.activeElement)) return;
+
+     // if there is some selected text, then start a new search for it
+    var selectedText = getSelectedText();
+    if (selectedText !== "") {
+      searchBox.value = selectedText;
+      searchBox.dispatchEvent(new Event('input'))
+      window.scrollTo({ top: 0, behavior: 'smooth' });
+      return; // let this event be handled by _handleKey for now, until we remove all this JS 
+    }
+
+    event.preventDefault();
+    searchBox.focus();
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  });
+
+  window.addEventListener('popstate', function (e) {
+    initStateFromQueryParams();
+  });
+
+  initStateFromQueryParams();
+  renderSearchHistory();
+  livePoll();
+}
 
 module.exports = {
-  Codesearch: Codesearch
+  init: init
 }
